refactor(application): use OmitIdAndTimestamps type in use case requests
