+++ src/application/usecases/create-question/create-question.usecase.ts
@@ -3,8 +3,9 @@ import type { QuestionsRepository } from '@/application/repositories/questions.r
 import type { QuestionProps } from '@/domain/entities/question/ports/question.props'
 import { Question } from '@/domain/entities/question/question.entity'
 import { QuestionWithTitleAlreadyRegisteredError } from './errors/question-with-title-already-registered.error'
+import type { OmitIdAndTimestamps } from '@/shared/types/custom/omit-id-and-timestamps'
 
-export type CreateQuestionRequest = QuestionProps
+export type CreateQuestionRequest = OmitIdAndTimestamps<QuestionProps>
 
 export class CreateQuestionUseCase implements UseCase {
   constructor (private readonly questionsRepository: QuestionsRepository) {}
